refactor(main):improved startup code
