+++ code/go/0chain.net/blobber/worker.go
@@ -1,0 +1,45 @@
+package main
+
+import (
+	"time"
+
+	"github.com/0chain/blobber/code/go/0chain.net/blobbercore/allocation"
+	"github.com/0chain/blobber/code/go/0chain.net/blobbercore/challenge"
+	"github.com/0chain/blobber/code/go/0chain.net/blobbercore/config"
+	"github.com/0chain/blobber/code/go/0chain.net/blobbercore/handler"
+	"github.com/0chain/blobber/code/go/0chain.net/blobbercore/readmarker"
+	"github.com/0chain/blobber/code/go/0chain.net/blobbercore/writemarker"
+	"github.com/0chain/blobber/code/go/0chain.net/core/common"
+	"github.com/spf13/viper"
+)
+
+func setupWorkers() {
+	var root = common.GetRootContext()
+	handler.SetupWorkers(root)
+	challenge.SetupWorkers(root)
+	readmarker.SetupWorkers(root)
+	writemarker.SetupWorkers(root)
+	allocation.StartUpdateWorker(root,
+		config.Configuration.UpdateAllocationsInterval)
+}
+
+func refreshPriceOnChain() {
+	var REPEAT_DELAY = 60 * 60 * time.Duration(viper.GetInt("price_worker_in_hours")) // 12 hours with default settings
+	for {
+		time.Sleep(REPEAT_DELAY * time.Second)
+		if err := registerBlobberOnChain(); err != nil {
+			continue // pass // required by linting
+		}
+	}
+}
+
+func keepAliveOnChain() {
+	const REPEAT_DELAY = 60 * 15 // 15 minutes
+
+	for {
+		time.Sleep(REPEAT_DELAY * time.Second)
+		if err := healthCheckOnChain(); err != nil {
+			continue // pass // required by linting
+		}
+	}
+}
